port to flowbox
This is an unifinished WIP commit.
The idea is to reimplement the ContentView widget using FlowBox
instead of IconView.
For now I am adding a Content2View widget so that I can leave the
old code untouched and only work on Alarm without touching World.
Once things are working we can rename back to ContentView and drop

The code is mostly working, the most noticeable TODO items are:
 - the selection checkbox looks ugly
 - port more css (font size, inactive and ringing states etc)
 - the way flowbox places tiles when resizing the windows looks bad,
   the space among tiles changes and this is especially noticeable
   because the second row moves when the action bar appears
 - once alarm is fine, port world (should be just be a matter of
   creating a custom Tile template)
 - once they are both ported delete a load of custom code (the custom
   iconview, the custom cell renderers, a lot of css etc).